01-Models/03-Embedded Models/1-embedding_openai.py
from langchain_openai import OpenAIEmbeddings
from dotenv import load_dotenv

load_dotenv()


emb = OpenAIEmbeddings(modeel="",dimentions=32)
result = emb.enbed_query("What is capital of india")

print(str(result))