01-Models/requirements.txt
# LangChain core
langchain
langchain-core

# OpenAI integration
langchain-openai
openai

# Anthropic (Claude) integration
langchain-anthropic
anthropic

# Google Gemini integration
langchain-google-genai
google-generativeai

# Hugging Face
langchain-huggingface
transformers
huggingface-hub

# Environment variables
python-dotenv

# Machine learning
numpy
scikit-learn
